backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
httpx==0.26.0
redis==5.0.1
celery==5.3.4
